create base activity
